requirements.txt
python-dotenv
pyopenssl
httpx[http2]
requests
rich